refactor(rp): extract eval endpoints into eval_routes.py
Move the 4 compare/eval endpoints (compare_message, select_candidate,
get_eval_set_detail, list_eval_sets) from routes.py into a dedicated
eval_routes.py module. Routes.py calls setup_eval_routes() to register
them. Removes ~170 lines from routes.py.
